+++ rust/exomonad-core/src/services/agent_control/cleanup.rs
@@ -3,6 +3,7 @@ use super::*;
 impl<
         C: super::super::HasGitHubClient
             + super::super::HasAcpRegistry
+            + super::super::HasOpencodeAcpRegistry
             + super::super::HasTeamRegistry
             + super::super::HasAgentResolver
             + super::super::HasProjectDir